Fixed fix for nanosleep
SVN revision: 5067

+++ src/system.c
@@ -3018,7 +3018,7 @@ INT ss_sleep(INT millisec)
 
    do {
       status = nanosleep(&ts, &ts);
-      if (ts.tv_sec < 0) 
+      if ((int)ts.tv_sec < 0) 
          break; // can be negative under OSX
    } while (status == -1 && errno == EINTR);
 #endif
